+++ tests-integration/tests/cpc_test/update.rs
@@ -17,6 +17,7 @@
 
 use datasketches::common::NumStdDev;
 use datasketches::cpc::CpcSketch;
+use datasketches::error::ErrorKind;
 use googletest::assert_that;
 use googletest::prelude::ge;
 use googletest::prelude::le;
@@ -59,4 +60,13 @@ fn test_many_values() {
     assert_that!(sketch.estimate(), ge(sketch.lower_bound(NumStdDev::One)));
     assert_that!(sketch.estimate(), le(sketch.upper_bound(NumStdDev::One)));
     assert!(sketch.validate());
+}
+
+#[test]
+fn test_max_serialized_bytes_validates_lg_k() {
+    assert!(CpcSketch::max_serialized_bytes(11).unwrap() > 0);
+    for lg_k in [3, 27] {
+        let error = CpcSketch::max_serialized_bytes(lg_k).unwrap_err();
+        assert_eq!(error.kind(), ErrorKind::InvalidArgument);
+    }
 }